renamed widgetReducers to widgets
