+++ Software/waveview/src/redux/reducers/index.tsx
@@ -1,8 +1,8 @@
 import { combineReducers } from 'redux';
-import horizontalWidgetReducer from './widgetReducers/horizontalWidgetReducer';
-import verticalWidgetReducer from './widgetReducers/verticalWidgetReducer';
-import measurementsWidgetReducer from './widgetReducers/measurementsWidgetReducer';
-import triggerWidgetReducer from './widgetReducers/triggerWidgetReducer';
+import horizontalWidgetReducer from './widgets/horizontalWidgetReducer';
+import verticalWidgetReducer from './widgets/verticalWidgetReducer';
+import measurementsWidgetReducer from './widgets/measurementsWidgetReducer';
+import triggerWidgetReducer from './widgets/triggerWidgetReducer';
 
 export default combineReducers(
   {
