+++ Software/waveview/src/redux/reducers/widgets/horizontalWidgetReducer.tsx
