+++ Software/waveview/src/redux/reducers/widgets/measurementsWidgetReducer.tsx
