+++ Software/waveview/src/redux/reducers/widgets/triggerWidgetReducer.tsx
