+++ Software/waveview/src/redux/reducers/widgets/verticalWidgetReducer.tsx
